refactor: extract shared categorical encoder mixin
Move the categorical feature/target encoding logic out of SklearnLikeModel into a reusable CategoricalEncoderMixin so tabular models share a single implementation instead of duplicating it.

Drop the per model CATEGORICAL_ENCODING strategy attribute and its enum: a column whose encoder preference is "label" is label encoded, everything else is one hot encoded. The previous fallback only affected columns with an unrecognized encoder preference, which does not occur in practice.
