+++ DashAI/back/models/scikit_learn/bayesian_ridge_regression.py
@@ -9,9 +9,6 @@ from DashAI.back.core.schema_fields import (
 )
 from DashAI.back.core.utils import MultilingualString
 from DashAI.back.models.regression_model import RegressionModel
-from DashAI.back.models.scikit_learn.sklearn_like_model import (
-    CategoricalEncodingStrategy,
-)
 from DashAI.back.models.scikit_learn.sklearn_like_regressor import SklearnLikeRegressor
 
 
@@ -219,7 +216,6 @@ class BayesianRidgeRegression(RegressionModel, SklearnLikeRegressor, _BayesianRi
     )
     COLOR: str = "#7E57C2"
     ICON: str = "Psychology"
-    CATEGORICAL_ENCODING = CategoricalEncodingStrategy.ONE_HOT
 
     def __init__(self, **kwargs) -> None:
         """Initialise the model by forwarding all kwargs to the parent class.
