+++ DashAI/back/models/scikit_learn/k_neighbors_regression.py
@@ -8,9 +8,6 @@ from DashAI.back.core.schema_fields import (
 )
 from DashAI.back.core.utils import MultilingualString
 from DashAI.back.models.regression_model import RegressionModel
-from DashAI.back.models.scikit_learn.sklearn_like_model import (
-    CategoricalEncodingStrategy,
-)
 from DashAI.back.models.scikit_learn.sklearn_like_regressor import SklearnLikeRegressor
 
 
@@ -197,7 +194,6 @@ class KNeighborsRegression(RegressionModel, SklearnLikeRegressor, _KNeighborsReg
     )
     COLOR: str = "#FFA726"
     ICON: str = "ScatterPlot"
-    CATEGORICAL_ENCODING = CategoricalEncodingStrategy.ONE_HOT
 
     def __init__(self, **kwargs) -> None:
         """Initialise the model by forwarding all kwargs to the parent class.
