+++ DashAI/back/models/scikit_learn/linear_regression.py
@@ -9,9 +9,6 @@ from DashAI.back.core.schema_fields import (
 )
 from DashAI.back.core.utils import MultilingualString
 from DashAI.back.models.regression_model import RegressionModel
-from DashAI.back.models.scikit_learn.sklearn_like_model import (
-    CategoricalEncodingStrategy,
-)
 from DashAI.back.models.scikit_learn.sklearn_like_regressor import SklearnLikeRegressor
 
 
@@ -160,8 +157,6 @@ class LinearRegression(RegressionModel, SklearnLikeRegressor, _LinearRegression)
     )
     COLOR: str = "#3F51B5"
     ICON: str = "ShowChart"
-
-    CATEGORICAL_ENCODING = CategoricalEncodingStrategy.ONE_HOT
 
     def __init__(self, **kwargs) -> None:
         """Initialise the model by forwarding all kwargs to the parent class.
