+++ DashAI/back/models/scikit_learn/ridge_regression.py
@@ -11,9 +11,6 @@ from DashAI.back.core.schema_fields import (
 )
 from DashAI.back.core.utils import MultilingualString
 from DashAI.back.models.regression_model import RegressionModel
-from DashAI.back.models.scikit_learn.sklearn_like_model import (
-    CategoricalEncodingStrategy,
-)
 from DashAI.back.models.scikit_learn.sklearn_like_regressor import SklearnLikeRegressor
 
 
@@ -279,8 +276,6 @@ class RidgeRegression(RegressionModel, SklearnLikeRegressor, _Ridge):
     )
     COLOR: str = "#2196F3"
     ICON: str = "ShowChart"
-
-    CATEGORICAL_ENCODING = CategoricalEncodingStrategy.ONE_HOT
 
     def __init__(self, **kwargs) -> None:
         """Initialise the model by forwarding all kwargs to the parent class.
